fix(conda): remove Python version tag
Prior to this commit, the `{{ python }}` tag was used to dynamically
specify the required Python based on the version of Python used during the
build.  However, since configuring the package to build as a `noarch:
python`, package, the Python version **MUST** not be specified or the
package will be locked to that version.

In this commit, remove the Python version tag, i.e., `{{ python }}` from
the build recipe requirements.